+++ src/pages/HomePage/pages/EventsManagement/pages/CreateEvent.tsx
@@ -1,6 +1,6 @@
-import { useAuth } from "@/auth/AuthWrapper";
 import axios, { pictureInstance as axiosPic } from "@/axiosInstance";
 import ImageUpload from "@/components/ImageUpload";
+import { useAuth } from "@/context/AuthWrapper";
 import UsePost from "@/CustomHooks/usePost";
 import usePut from "@/CustomHooks/usePut";
 import api from "@/utils/apiCalls";
